color: consolidate HSL, OKLab, OKLCH math into scene/utils/color/colors.ts
Three previously-separate modules (hsl.ts, oklch.ts, colorInterp.ts)
covered overlapping color-math territory. Merge into one colors.ts
with sections for each space:

- HSL string helpers (mirror hsl.glsl for shader parity)
- OKLab Cartesian conversion (out-params for hot paths)
- OKLCH polar wrapper (allocates; palette-generation use cases)
- linear sRGB → hex encoding
- interpolateOklch (perceptual color interpolation)

Update the three importers (authorColor.ts, treeRenderer.ts, gem.ts —
gem in next commit) and relocate colorInterp.test.ts next to its
module. Delete the old hsl.ts and colorInterp.ts.

The oklchToLinearRgb function previously inlined in authorColor.ts
is gone — it now lives in the shared module and reuses oklabToLinear
internally, removing the duplicate OKLab→linear-sRGB matrix.
