+++ app/tests/scene/utils/color/colorInterp.test.ts
